Fix #44, #54, #56: expand_file(): tilde (~) for home directory does not work in windows
Delegate to Path::Tiny the task of interpreting a '~' as the first char in a path name passed to expand_file().
This solves issue #44 (expand_file(): tilde (~) for home directory does not work in windows).

Issues #54 and #56 do not occur in my platform, but are reported by CPAN Tests failures.
I think they are solved by closing #44, but I need to upload a developer release to CPAN and wait for test reports
to be sure.