Organize tests
Update test files for the new implementation so that all tests pass.

With the new implementation each macro expansion is evaluated once and
not looked-in again - that means that the output is not rescanned for macros,
which causes some of the previous tests to fail.

There are some differences in the output of list_... to be more consistent -
always exactly one newline at the end of each definition.

Eval error messages are now correctly captured instead of being sent to
STDERR.

The error message for non-matched structure is now the same for all structures.

No need to check for nested structures in %DEFINE, as now the module collects
all text up to %END_DEFINE and evaluates it on call time, delaying any
error message to then.

Different behavior in %CASE, as the text is collected up to %END_CASE or
the next %CASE and output in one chunk, instead of outputing line by line.

Undefining a not-defined entity does no longer output an error message.

Remove diag messages for fixed bugs.